Extract actions to seperate file.
